package refactoring for SoC
