+++ cli/src/main/java/com/devonfw/tools/ide/tool/custom/CustomToolJson.java
@@ -1,4 +1,4 @@
-package com.devonfw.tools.ide.tool.repository;
+package com.devonfw.tools.ide.tool.custom;
 
 import com.devonfw.tools.ide.os.OperatingSystem;
