+++ cli/src/main/java/com/devonfw/tools/ide/tool/custom/CustomToolRepository.java
@@ -1,6 +1,8 @@
-package com.devonfw.tools.ide.tool.repository;
+package com.devonfw.tools.ide.tool.custom;
 
 import java.util.Collection;
+
+import com.devonfw.tools.ide.tool.repository.ToolRepository;
 
 /**
  * Interface for the custom {@link ToolRepository}.
