+++ cli/src/main/java/com/devonfw/tools/ide/tool/mvn/MvnArtifact.java
@@ -4,7 +4,6 @@ import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import com.devonfw.tools.ide.tool.repository.MvnRepository;
 import com.devonfw.tools.ide.tool.repository.SoftwareArtifact;
 
 /**
