+++ cli/src/main/java/com/devonfw/tools/ide/tool/pip/PipRepository.java
@@ -1,14 +1,12 @@
-package com.devonfw.tools.ide.tool.repository;
+package com.devonfw.tools.ide.tool.pip;
 
 import java.util.List;
 
 import com.devonfw.tools.ide.context.IdeContext;
 import com.devonfw.tools.ide.json.JsonMapping;
 import com.devonfw.tools.ide.tool.ToolCommandlet;
-import com.devonfw.tools.ide.tool.pip.Pip;
-import com.devonfw.tools.ide.tool.pip.PipArtifact;
-import com.devonfw.tools.ide.tool.pip.PipArtifactMetadata;
-import com.devonfw.tools.ide.tool.pip.PypiObject;
+import com.devonfw.tools.ide.tool.repository.AbstractToolRepository;
+import com.devonfw.tools.ide.tool.repository.ArtifactToolRepository;
 import com.devonfw.tools.ide.version.VersionIdentifier;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
