+++ cli/src/test/java/com/devonfw/tools/ide/context/PipRepositoryMock.java
@@ -11,7 +11,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.stream.Stream;
 
-import com.devonfw.tools.ide.tool.repository.PipRepository;
+import com.devonfw.tools.ide.tool.pip.PipRepository;
 import com.github.tomakehurst.wiremock.junit5.WireMockRuntimeInfo;
 
 /**
